+++ src/models/weapon/Weapon.ts
@@ -1,6 +1,6 @@
 import { JsonReader, JsonObject } from "config_file.js";
 import EnkaClient from "../../client/EnkaClient";
-import StatusProperty, { FightProp } from "../StatusProperty";
+import StatProperty, { FightProp } from "../StatProperty";
 import WeaponData from "./WeaponData";
 import WeaponRefinement from "./WeaponRefinement";
 
@@ -25,7 +25,7 @@ class Weapon {
     /**  */
     readonly isAwaken: boolean;
     /**  */
-    readonly weaponStats: StatusProperty[];
+    readonly weaponStats: StatProperty[];
 
     readonly _data: JsonObject;
 
@@ -58,7 +58,7 @@ class Weapon {
 
         this.isAwaken = this.ascension >= 2;
 
-        this.weaponStats = json.get("flat", "weaponStats").mapArray((_, p) => new StatusProperty(p.getAsString("appendPropId") as FightProp, p.getAsNumber("statValue"), enka, true));
+        this.weaponStats = json.get("flat", "weaponStats").mapArray((_, p) => new StatProperty(p.getAsString("appendPropId") as FightProp, p.getAsNumber("statValue"), enka, true));
 
     }
 }